requirements.txt
pyspark>=3.4.0
great-expectations>=1.8.0
apache-airflow>=2.8.0
pytest>=7.0.0
